// src/app/core/models/menum.ts
// tslint:disable-next-line: class-name
export class _Menum {
    menu_pkid: string;
    menu_code: string;
    menu_name: string;
    menu_route1: string;
    menu_route2: string;
    menu_type: string;
    menu_order: number;

    menu_sep: boolean;
    menu_module_id: string;
    menu_module_name: string;
    menu_module_order: number;
    rec_mode: string;
}


// tslint:disable-next-line: class-name
export class User_Menu {
    menu_pkid: string;
    rights_pkid: string ;
    rights_au_id: string ;
    menu_module_id: string ;
    module_order: number;
    module_name: string ;
    menu_name: string ;
    menu_param: string ;
    menu_locked: string ;
    menu_hiden: string ;
    menu_order: number;
    menu_xap_name: string ;
    menu_xap_dll: string ;
    menu_xap_class: string ;
    rights_company: string ;
    rights_admin: string ;
    rights_add: string ;
    rights_edit: string ;
    rights_view: string ;
    rights_delete: string ;
    rights_print: string ;
    rights_email: string ;
    menu_group_code: string ;
    menu_group_name: string ;
    menu_display_name: string ;
    menu_checked: boolean;
    menu_route1: string;
    menu_route2: string;
}